Moved test data to test module.
